Extract HomeController generator
