+++ lib/generators/shopify_app/home_controller/templates/home_controller.rb
