+++ lib/generators/shopify_app/home_controller/templates/index.html.erb
