+++ lib/generators/shopify_app/home_controller/templates/shopify_app_ready_script.html.erb
