+++ lib/generators/shopify_app/shopify_app_generator.rb
@@ -9,6 +9,7 @@ module ShopifyApp
 
       def run_all_generators
         generate "shopify_app:install #{@opts.join(' ')}"
+        generate "shopify_app:home_controller"
         generate "shopify_app:shop_model"
 
         generate "shopify_app:controllers"
